🔨 refactor(cli): Separate each command into a separate file
