+++ crates/cli/src/commands/add.rs
@@ -3,22 +3,6 @@ use futures::future::join_all;
 
 use crate::parser::Package;
 
-pub fn install_command() {
-    todo!("Install dependencies");
-}
-
-pub fn update_command() {
-    todo!("Update dependencies");
-}
-
-pub fn run_command(_script: String) {
-    todo!("Run a script");
-}
-
-pub fn list_command() {
-    todo!("List dependencies");
-}
-
 pub async fn add_command(packages: Vec<String>, _dev: bool, _peer: bool, _optional: bool) {
     let client = create_client();
 
@@ -59,25 +43,8 @@ pub async fn add_command(packages: Vec<String>, _dev: bool, _peer: bool, _option
                 let package: NpmPackage = res.json().await.unwrap();
 
                 println!("Package: {}", package.name);
-            },
+            }
             Err(err) => eprintln!("Error: {}", err),
         }
     }
-}
-
-pub fn remove_command(package: String) {
-    let package = Package::parse(&package);
-
-    match package {
-        Ok(package) => {
-            println!("{:?}", package);
-        }
-        Err(e) => {
-            println!("{}", e);
-        }
-    }
-}
-
-pub fn init_command(_name: Option<String>) {
-    todo!("Create a new project");
 }
