+++ crates/cli/src/commands/init.rs
@@ -1,0 +1,3 @@
+pub fn init_command(_name: Option<String>) {
+    todo!("Create a new project");
+}
