+++ crates/cli/src/commands/install.rs
@@ -1,0 +1,3 @@
+pub fn install_command() {
+    todo!("Install dependencies");
+}
