+++ crates/cli/src/commands/list.rs
@@ -1,0 +1,3 @@
+pub fn list_command() {
+    todo!("List dependencies");
+}
