+++ crates/cli/src/commands/mod.rs
@@ -1,0 +1,15 @@
+mod add;
+mod init;
+mod install;
+mod list;
+mod remove;
+mod run;
+mod update;
+
+pub use add::add_command;
+pub use init::init_command;
+pub use install::install_command;
+pub use list::list_command;
+pub use remove::remove_command;
+pub use run::run_command;
+pub use update::update_command;
