+++ crates/cli/src/commands/run.rs
@@ -1,0 +1,3 @@
+pub fn run_command(_script: String) {
+    todo!("Run a script");
+}
