+++ crates/cli/src/commands/update.rs
@@ -1,0 +1,3 @@
+pub fn update_command() {
+    todo!("Update dependencies");
+}
